refactor(fpss): extract delta.rs
Move DeltaState, the STOP_SUPPRESS_DURATION constant, and the per-tick
field-count constants (QUOTE_FIELDS/TRADE_FIELDS/OI_FIELDS/OHLCVC_FIELDS)
out of fpss/mod.rs. Fields and methods touched by decode_frame (ohlcvc,
last_was_date, last_stop, decode_tick, clear, is_in_stop_suppression_window)
are exposed as pub(super). Decode-tick tests still live in mod.rs and move
alongside decode_frame in a later commit.
